change provider asset route urls and directories
