+++ admin/templates/osf/providerassetfile_form.html
@@ -8,7 +8,7 @@
 <div class="row">
     <div class="col-md-12">
         {% if perms.osf.delete_providerassetfile %}
-            <a class="btn btn-danger" href={% url 'asset_files:delete' object.id %}>Delete asset</a>
+            <a class="btn btn-danger" href={% url 'provider_asset_files:delete' object.id %}>Delete asset</a>
         {% endif %}
     </div>
 </div>
